+++ lib/doggo.ex
@@ -2784,6 +2784,10 @@ defmodule Doggo do
   attr :open, :boolean, default: false, doc: "Initializes the modal as open."
   attr :on_cancel, JS, default: %JS{}
 
+  attr :close_label, :string,
+    default: "Close",
+    doc: "Aria label for the close button."
+
   slot :title, required: true
   slot :inner_block, required: true, doc: "The modal body."
 
@@ -2823,7 +2827,7 @@ defmodule Doggo do
             <.link
               href="#"
               class="modal-close"
-              aria-label="Close"
+              aria-label={@close_label}
               phx-click={JS.exec("data-cancel", to: "##{@id}")}
             >
               <%= render_slot(@close) %>